requirements.txt
Flask==3.1.0
RPi.GPIO==0.7.1
mfrc522==0.0.7
pandas==2.2.3